refactor: check download size before cloning (#122)
Currently the download directory size is checked in the `crawler`. This
doesn't make a whole lot of sense, since the crawler can schedule tens
of repos to be cloned and then the size of the download dir can greatly
exceed the set maximum. With this changes repos won't be cloned if
there's not enough space and instead will be retried periodically.

Keep in mind that this still has some faulty logic, due to
implementation of `_clone_repo` method. There's a path that says `f"Repo
{repo.git_url} already exists, fetched from disk"` but I don't actually
understand why would there be a repository downloaded already if we
trigger `process_repo` only once per.