refactor: Sass migration (#227)
I migrated Sass (the CSS extension) to a newer version that uses the
updated file import rules and introduces namespaces. The old version was
deprecated and generated a lot of warnings in the console.
Due to warnings, I also updated the usage of the `end` option to the
recommended `flex-end`.